pyproject.toml
[tool.poetry]
name = "foodify"
version = "0.1.0"
description = ""
authors = ["Viktor Nikel <v.nikell.28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pydantic = "^1.10.8"
python-dotenv = "^1.0.0"
openai = "^0.27.7"
loguru = "^0.7.0"
python-telegram-bot = "^20.3"
telegram-menu = "^2.1.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
asyncpg = "^0.27.0"
greenlet = "^2.0.2"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.1"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.3.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = 3.11

[tool.flake8]
max-line-length = 88